Move function to blockchain
